refactor: 애플리케이션 레이어 DTO 및 Command 구조 리팩토링
