+++ src/main/kotlin/kr/hhplus/be/application/balance/BalanceCommand.kt
@@ -1,5 +1,10 @@
 package kr.hhplus.be.application.balance
 
+data class BalanceChargeCommand(
+    val userId: Long,
+    val amount: Int,
+)
+
 data class BalanceDeductCommand(
     val userId: Long,
     val amount: Int
