--- src/main/kotlin/kr/hhplus/be/application/product/ProductRankingDto.kt
@@ -1,7 +1,0 @@
-package kr.hhplus.be.application.product
-
-data class ProductRankingInfo(
-    val productId: Long,
-    val productName: String,
-    val rank: Int,
-)
