+++ src/main/kotlin/kr/hhplus/be/application/service/ProductRankingService.kt
@@ -1,6 +1,6 @@
 package kr.hhplus.be.application.service
 
-import kr.hhplus.be.application.product.ProductRankingInfo
+import kr.hhplus.be.application.product.ProductDto.ProductRankingInfo
 import kr.hhplus.be.domain.product.ProductRankingRepository
 import org.springframework.stereotype.Service
