+++ src/main/kotlin/kr/hhplus/be/domain/product/ProductRankingRepository.kt
@@ -1,6 +1,6 @@
 package kr.hhplus.be.domain.product
 
-import kr.hhplus.be.application.product.ProductRankingInfo
+import kr.hhplus.be.application.product.ProductDto.ProductRankingInfo
 
 interface ProductRankingRepository {
     fun findTopProducts(): List<ProductRankingInfo>
