+++ src/main/kotlin/kr/hhplus/be/infrastructure/persistence/ProductRankingRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package kr.hhplus.be.infrastructure.persistence
 
-import kr.hhplus.be.application.product.ProductRankingInfo
+import kr.hhplus.be.application.product.ProductDto.ProductRankingInfo
 import kr.hhplus.be.domain.product.ProductRankingRepository
 import kr.hhplus.be.infrastructure.persistence.repository.ProductJpaRepository
 import org.springframework.stereotype.Component
